// src/app/Models/flight.ts
export class Flight {
    flightNumber: number;
    intrument_Used:string; 
    from_Place: string;  
    to_Place:string;  
    airLine:string;  
    startDateTime:Date;  
    endDateTime: Date;  
    totalBusinessClassSeats: number;  
    totalNonBusinessClassSeats: number;  
    noOfRows: number; 
    mealType: string;
    scheduledDays: string; 
    price: number; 
}